requirements.txt
# pip install -r requirements.txt
fastapi
uvicorn
pywebio
python-dotenv
cloudflare
proxmoxer
requests
guacapy